fix network error on test
